interior/backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
mysqlclient>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

